refactor: typewriterfactory to typewriters
BREAKING CHANGE: The TypeWriterFactory is now called TypeWriters
