+++ packages/generator/src/index.ts
@@ -2,6 +2,7 @@ export { default as cli } from './cli'
 export { default as Generator } from './Generator'
 export * from './runtypes'
 export * from './TypeWriter'
+export { default as TypeWriters } from './TypeWriters'
 export {
   escapeQuottedPropName,
   getGenerics,
